+++ web/CHANGELOG.md
@@ -3,6 +3,10 @@
 - Add Share to PDP [#963](https://github.com/mobify/platform-scaffold/pull/963)
 - Fix account dashboard path [#986](https://github.com/mobify/platform-scaffold/pull/986)
 - Fetch shipping methods upon shipping address form completion [#990](https://github.com/mobify/platform-scaffold/pull/990)
+- Added Offline mode analytics tracking [#988](https://github.com/mobify/platform-scaffold/pull/988)
+    - Dispatch actions to store data for offline mode
+    - Dispatch actions to send analytics data about offline mode session
+    - Remove template higher order component from offline container
 - Fix bug with account login/logout for Merlin's connector [#992](https://github.com/mobify/platform-scaffold/pull/992)
 - Add lock/unlock scroll when open/closing modals [#989](https://github.com/mobify/platform-scaffold/pull/989)